// src/app/models/news.ts
export class News {
    id!: string;
    haberId!: string;
    haberBasligi?: string;
    haberYazan?: string;
    katId?: string;
    categoryId?: string;
    haberManseti?: string;
    haberYazisi?: string;
    haberKayitTarihi?: string;
    haberDuzenlemeTarihi?: string;
    haberResim?: string;
    haberTiklanma?: number;
}